requirements.txt
requests>=2.31.0,<3
urllib3>=2.0.0,<3
python-dotenv>=1.0.0,<2
